backend/requirements.txt
# System dependency: ffmpeg is required for audio transcoding but is not a Python package.
# Install ffmpeg separately on your machine before running the backend.
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
celery[redis]==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
alembic==1.13.3
twilio==9.3.6
aiofiles==24.1.0
pydantic[email]
deep-translator==1.11.4